+++ helpers/src/directIOHelper.h
@@ -25,6 +25,10 @@
 namespace one {
 namespace helpers {
 
+constexpr auto DIRECT_IO_HELPER_UID_ARG = "uid";
+constexpr auto DIRECT_IO_HELPER_GID_ARG = "gid";
+constexpr auto DIRECT_IO_HELPER_PATH_ARG = "root_path";
+
 /**
 * The PosixHelperCTX class represents context for all POSIX compliant helpers
 * and its object is passed to all helper functions.
@@ -117,7 +121,7 @@ public:
     void ash_truncate(
         CTXPtr ctx, const boost::filesystem::path &p, off_t size, VoidCallback);
 
-    void ash_open(CTXPtr ctx, const boost::filesystem::path &p, FlagsSet flags,
+    void ash_open(CTXPtr ctx, const boost::filesystem::path &p, int flags,
         GeneralCallback<int>);
     void ash_read(CTXPtr ctx, const boost::filesystem::path &p,
         asio::mutable_buffer buf, off_t offset, const std::string &fileUuid,