+++ include/options.h
@@ -182,7 +182,7 @@ private:
     DECL_CMDLINE_SWITCH_DEF(debug_gsi, "", false, "enable GSI debug output")
     DECL_CMDLINE_SWITCH_DEF(help, ",h", false, "print help")
     DECL_CMDLINE_SWITCH_DEF(version, ",V", false, "print version")
-    DECL_CMDLINE_SWITCH_DEF(directio, "", false, "enable DirectIO")
+    DECL_CMDLINE_SWITCH_DEF(proxyio, "", false, "force ProxyIO")
     DECL_CONFIG_DESC(config, std::string, "path to user config file")
     DECL_CONFIG_DEF(enable_env_option_override, bool, true)
     /* clang-format on */