+++ src/cache/helpersCache.h
@@ -12,6 +12,8 @@
 #include "communication/communicator.h"
 #include "helpers/IStorageHelper.h"
 #include "helpers/storageHelperFactory.h"
+#include "scheduler.h"
+#include "storageAccessManager.h"
 
 #include <asio/io_service.hpp>
 #include <asio/executor_work.hpp>
@@ -22,6 +24,11 @@
 #include <utility>
 
 namespace one {
+namespace messages {
+namespace fuse {
+class StorageTestFile;
+}
+}
 namespace client {
 
 /**
@@ -33,15 +40,21 @@ public:
     using HelperPtr = std::shared_ptr<helpers::IStorageHelper>;
 
 private:
-    communication::Communicator &m_communicator;
-    asio::io_service m_ioService{1};
-    asio::executor_work<asio::io_service::executor_type> m_work =
-        asio::make_work(m_ioService);
+    void requestStorageTestFileCreation(
+        const std::string &fileUuid, const std::string &storageId);
 
-    std::thread m_thread;
+    void handleStorageTestFile(
+        std::shared_ptr<messages::fuse::StorageTestFile> testFile,
+        const std::string &storageId, unsigned int attempts);
 
-    helpers::StorageHelperFactory m_helperFactory{
-        m_ioService, m_ioService, m_ioService, m_communicator};
+    void requestStorageTestFileVerification(
+        const messages::fuse::StorageTestFile &testFile,
+        const std::string &storageId, std::string fileContent);
+
+    void handleStorageTestFileVerification(
+        const std::error_code &ec, const std::string &storageId);
+
+    enum class AccessType { DIRECT, PROXY };
 
     struct HashCompare {
         bool equal(const std::tuple<std::string, bool> &j,
@@ -49,12 +62,25 @@ private:
         size_t hash(const std::tuple<std::string, bool> &k) const;
     };
 
-    tbb::concurrent_hash_map<std::tuple<std::string, bool>,
-        HelperPtr, HashCompare> m_cache;
+    communication::Communicator &m_communicator;
+    Scheduler &m_scheduler;
+    asio::io_service m_ioService{1};
+    asio::executor_work<asio::io_service::executor_type> m_work =
+        asio::make_work(m_ioService);
+    std::thread m_thread;
+    helpers::StorageHelperFactory m_helperFactory{
+        m_ioService, m_ioService, m_ioService, m_communicator};
+    StorageAccessManager m_storageAccessManager;
+
+    tbb::concurrent_hash_map<std::tuple<std::string, bool>, HelperPtr,
+        HashCompare> m_cache;
+    tbb::concurrent_hash_map<std::string, AccessType> m_accessType;
 
 public:
-    using ConstAccessor = decltype(m_cache)::const_accessor;
-    using Accessor = decltype(m_cache)::accessor;
+    using ConstAccessTypeAccessor = decltype(m_accessType)::const_accessor;
+    using AccessTypeAccessor = decltype(m_accessType)::accessor;
+    using ConstCacheAccessor = decltype(m_cache)::const_accessor;
+    using CacheAccessor = decltype(m_cache)::accessor;
 
     /**
      * Constructor.
@@ -63,7 +89,8 @@ public:
      * @param communicator Communicator instance used to fetch helper
      * parameters.
      */
-    HelpersCache(communication::Communicator &communicator);
+    HelpersCache(
+        communication::Communicator &communicator, Scheduler &scheduler);
 
     /**
      * Destructor.
@@ -73,13 +100,13 @@ public:
 
     /**
      * Retrieves a helper instance.
-     * @param spaceId Space id for which to retrieve a helper.
+     * @param fileUuid UUID of a file for which helper will be used.
      * @param storageId Storage id for which to retrieve a helper.
-     * @param forceClusterProxy Determines whether to return a ClusterProxy
-     * helper.
+     * @param forceProxyIO Determines whether to return a ProxyIO helper.
      * @return Retrieved helper instance.
      */
-    HelperPtr get(const std::string &storageId, const bool forceClusterProxy = false);
+    HelperPtr get(const std::string &fileUuid, const std::string &storageId,
+        bool forceProxyIO = false);
 };
 
 } // namespace one